[Mono]: Fix Mono profiler EventPipe provider instrumentation feature. (#110618)
In .NET 8 the mono profiler EventPipe provider was disabled by default.
Part of that change, the creation of Mono profiler instance was moved
but call to setup call instrumentation could do an early call to setup
callback using a profiler instance that has not yet been allocated.

Fix move call to a point where the provider has been created. Fix also
makes sure that providing a callspec at startup make sure the call
instrumentation callback lives for the lifetime of the process.