Spill to both parameter and new mappings
For GC refs in structs these have to be zeroed anyway, so we might as
well just spill to the original parameter as well as the new mapping.
This means we can avoid the manually inserted spill in the init block,
and we retain the property that parameters are fully defined by the
prolog.