Fix LoaderAllocator of DynamicMethodTable
There is one `DynamicMethodTable` allocated per `Module` if needed. When its
instance is being created, an incorrect `LoaderAllocator` is passed to the
`CreateMinimalMethodTable` function. Instead of the `Module`'s
`LoaderAllocator`, it passes in the default one.
That fires an assert in non-release builds of the runtime and it results in
incorrect marking of the `MethodTable` as not collectible in case the `Module`'s
LoaderAllocator is collectible.

This change fixes it by passing in the `LoaderAllocator` of the module
instead.

Close #110987