Fix WriteNamedBitList for enums with negative values.
WriteNamedBitList internally operates on a ulong when writing down bits. If the enum is backed by a signed type and the enum value is negative, it would be signed extended when being converted to a long, then ulong. This resulted in all negative values being written down as 64 bits set regardless of the width of the enum type.

This changes the write to instead convert to the signed types corresponding unsigned type first, then converting to a ulong, which results in the value being zero extended.