JIT: split loop headers that are also try entries (#110880)
when there are backedges that have exited the try. This allows
the preheader and header to be in the same EH region and may
unblock other optimizations.

This is done by splitting the header block and making the trailing
portion be the try entry. Perform a mid (or end) block split if the
initial (or all) statements in the header cannot throw.

Fixes #96887.

Enable cloning of loops with EH.