Change metadata handle split to 7-bit handle type
Fixes dotnet/runtimelab#1581

This allows addressing up to 32 MB of metadata. This should cover all known cases where people hit this in the past.

Granted, none of the user hits are for scenarios where it would be advisable to use native AOT. If one needs to reflection-root 60+ MB of IL, it's inadvisable to use native AOT or trim in general since the possible (small) savings will never outweigh (huge) reliability/predictabillity loss caused by the 60 MB of trim unsafe code.)