Another attempt at fixing #99198
The problem was always `// Note: We don't set the IsUnboxingStub flag on template methods (all template lookups performed at runtime are performed with this flag not set`, I tried working around it in the original fix, but looks like I actually need the function pointer in #111178 (the tests are not failing but I have a local test that does).

So trying an alternative approach that just deletes the weird code. It's possible this only had to be weird due to universal shared code.