[mono][wasm] AOT more methods containing EH clauses
WASM aot supports the `deopt` mode for a method, in which case the method will be aot compiled but the exception handlers are later invoked via the interpreter, if an exception is thrown. This is an alternative to fully interpreting the method in question. A problem can arise if an exception is caught inside a finally handler, because the start of the finally handler will be executed with AOT and then, once the exception is caught, the rest of the handler would be executed with the interprer. This is because the implementation of ENDFINALLY in the interpreter requires some information to have been initialized when the handler is called, initialization which is missing since the handler was called by AOT.

The original code referenced this issue, but the condition to detect this situation was very relaxed, including common cases like `try { try {} catch {} } finally { }`. This commit changes the check so it is only hit for scenarios like `try { } finally { try { } catch { } }`.