JIT: Switch `DataFlow::ForwardAnalysis` to be RPO based
`DataFlow::ForwardAnalysis` used a worklist based algorithm where it
kept visiting successor blocks as long as information was being
propagated until closure. However, this successor-based visit was using
regular successors only, which does not account for EH second-pass flow.

This was not generally a correctness problem as the only users of this
class are CSE and assertion prop, and neither of them have a correctness
requirement of propagating facts along filter -> finally/fault edges,
but the imprecision in the modelling could result in disagreements in
reachability with the DFS visit.

Concretely we could end up with CSE's availability analysis finding that
a nested finally/fault block was unreachable (because its corresponding
try-entry was unreachable), while the DFS visit considered it reachable
due to a spurious filter predecessor. This mismatch meant that the
handler was never visited by the CSE dataflow which would result in all
CSEs in some of the handler blocks being marked as available. We would
then end up creating a use in one of these blocks, which SSA updating
would not be able to find a corresponding def for.

Fix the situation by switching `DataFlow::ForwardAnalysis` to compute a
DFS tree and computing the closure based on its RPO.

Fix #111345