JIT: move second pass of EH opts earlier
to avoid messing up loop exit canonicalization (which is sensitive
to loops with internal EH flow).

Fixed #110959.

Also, remove some of the ceremony around `fgModified` -- seems like this
is no longer useful/needed.

Defer deciding if we need a PSPSym until we create funclets. This removes
unused PSPSym allocations in some cases. Also, there is no benefit to
tracking the PSPSym, so stop doing that.