Disable polyrec test on native AOT (#111333)
This has a deep generic recursion that generates throwing code. It also uses threads so the unhandled exception crashes the merged test runner instead of being reported as a failure. The reporting at the end of test run then reports 0 failing test. I assume this would still be failing in Helix, so it's not a big deal, but I missed a couple failing tests from the Loader tree that I need to bucket.