JIT: don't block tail calls for unreferenced exposed locals
Ignore unreferenced address exposed locals when determining if it's safe to
make an implicit tail call. We consult the RCS_EARLY ref count for this;
hopefully it's conservatively correct.

This fixes some cases that come up if we stack allocate and then dead code
a new array, as well as other cases we're currently missing.