Mark JsonNode.WriteTo as obsolete with SYSLIB0060
This commit introduces the diagnostic ID `SYSLIB0060` for the obsolescence of the `JsonNode.WriteTo` method. Users are advised to use `JsonSerializer.Serialize{Async}` instead.

The new diagnostic message and ID have been added to the `Obsoletions` class, and the documentation in `list-of-diagnostics.md` has been updated accordingly.

The `[Obsolete]` attribute has been applied to the `WriteTo` method in `JsonArray`, `JsonNode`, `JsonObject`, and `JsonValue`.

Additionally, warnings related to the use of the obsolete method have been suppressed in various locations.