JIT: fix loop cloning when loop's last block is within a hander
In some unusual cases the last block of a loop might be in the middle of a
handler region (say the loop backedge is in a catch, and the catch has
alternate flow that exits the loop).

Make sure that we properly place the slow loop preheader into the same EH regions
as the fast loop preheader.

Fixes #111344