JIT: Cache start of cold section for reuse in 3-opt layout (#111365)
During cold code motion, if fgMoveColdBlocks finds a call-finally pair that begins with a cold block, it will try to move the entire pair, regardless of whether the tail is cold or not. If the tail isn't cold, 3-opt layout's detection of the last hot block may terminate early, causing the cold part of an EH region to also be considered for reordering. Because 3-opt expects each EH subregion being reordered to be contiguous, this split causes asserts. Caching the cold region start point in fgMoveColdBlocks means we don't need to find the first cold block in 3-opt, thus removing the potential of getting it wrong.

While I was here, I decided to enhance fgMoveColdBlocks so that it can move cold call-finally pairs at the end of the hot main method body, and clean up 3-opt's initialization logic to reduce its allocation sizes. Fixes #111207.