Skip Init and Static helpers when looking for a caller (#111446)
* Skip Init and Static helpers when looking for a caller

We have learned that real world code may call Assembly.GetCallingAssembly in a static constructor. It returns different value now that there is an extra managed frame introduced by HMF removal. The fix adds the managed helper types to the long ad-hoc list of types that should be skipped for this and similar APIs. It does not change the fact that Assembly.GetCallingAssembly is a slow fragile unreliable API that should not be used.

Fixes #111399

* Add regression test