Fix stack overflow in compiler-generated state (#109207)
When building the compiler-generated type info for a type, there are
some cases that produce warnings due to unexpected IL. This can
recurse back into the compiler-generated type info logic when checking
for warning suppressions, leading to stack overflow.

This fixes the issue by delaying the logging of warnings until after
the compiler-generated type info has been fully built for a given
type.