Fix explicit offset of ByRefLike fields.
A ByRefLike type was previously forced to be
pointer aligned so ref fields were implicitly aligned.
Using Explicit layout one can create valid aligned
ByRefLike types with no object pointers. This change
propagates the base offset of the ByRefLike type and
includes that for determining offsets of fields in the
ByRefLike type.