JIT: Relax assertion for handlers of unreachable try clauses
For unreachable try clauses it is possible to still see spurious flow
into their handlers because of second pass EH. We can then end up with a
case where a handler is dominated by one of its enclosed blocks. For
definitions in that enclosed block we thus won't find any phi in the
enclosing handler; tolerate this case.

Fox #111341