JIT: Don't let 4-opt layout modify other EH regions (#111637)
When trying to create fallthrough for a jump that begins after an EH region, and ends before the EH region, don't consider a split point within the EH region; otherwise, the swap will break the invariant that EH entries are not moved relative to other blocks in the region. Fixes #111347.