JIT: Fix embedded broadcast simd size (#111638)
This ensures that when a constant vector is replaced with a broadcast node, the replacement node is of the correct size.  Since only the scalar is read from the broadcast node, SIMD size doesn't really matter for codegen, but this fix satisfies the assert in #111613 by ensuring we replace like for like.