Add FrozenDictionary specialization for integers / enums
This adds a specialized dictionary implementation for when TKey is an integer (some integer types) or an enum backed by an integer. It employs an array that can be indexed into directly, handling both the cases where the values are contiguous from zero and where they're either non-contiguous or not from zero. A density threshold is used to decide when to fallback to another implementation rather than expending more memory on unused array slots.