Removing Entry2MethodDesc as it is unnecessary (#111756)
It appears that the pathways that Entry2MethodDesc has for virtual resolution are unnecessary. Testing has shown that everything seems to work without this logic, and removing excess baggage is always good, and a bit less code should run a bit faster. As a bonus, this simplifies the logic around adding support for cached interface dispatch in CoreCLR.

Also, fix the crossgen2 test legs, as they are currently broken due to issues with the now NativeAOT'd crossgen2 binary.